funcs: convert comment format using _scripts_/docuriumify.sh
Note, the only difference between the current master branch at
Predelnik/notes is that "//" is not replaced with "///" if it is
in the middle of the line.

This effectively solves the problem with "https://" becoming
"https:///"; e.g.

-///    * https:///en.wikipedia.org/wiki/Linear_congruential_generator#Parameters_in_common_use
+///    * https://en.wikipedia.org/wiki/Linear_congruential_generator#Parameters_in_common_use

It also preserves the format of comments in pseduo-code.
If we should keep pseudo-code in comments at all, can be
further discussed in the future. For now, lets just keep
them as is.

 ///    switch {
 ///    case loot_id > 0:
-///       /// regular item drop.
+///       // regular item drop.
 ///       item_id = loot_id - 1
 ///    case loot_id < 0:
-///       /// unique item drop.
+///       // unique item drop.
 ///       unique_item_id = -loot_id - 1
 ///    case loot_id == 0:
-///       /// no item drop.
+///       // no item drop.
 ///    }