enums: convert comment format using _scripts_/docuriumify.sh
Preserve https:// and TODO comments, otherwise identical to
the master branch in Predelnik/notes with regards to enum
member comments.

Two additions of Predelnik/notes have not been considered by
the script and will be added by hand later. The addition of
"bitflag" comments for docurium and the addition of the include
header guard ENUMS_H.