rdata: use PSX names for globals of gmenu.h
Note, the lfontframe and lfontkern globals don't have
PSX debug info, however, they correspond to the
fontframe and fontkern variables of control.h, but
apply to large instead of small fonts. Thus the names
are derived as such. Also, the PSX debug info does
include global variables of the format FontTab,
LFontTab, MFontTab. As such, M and L prefixes are used
for medium and large fonts, and a missing prefix is
used for small fonts.