rdata: use PSX names for globals of interfac.h
Rename progress_bar_colours and progress_bar_screen_pos
to BarColor and BarPos, respectively.

Note, the BarColor and BarPos globals don't have
PSX debug info. However, the progress bar is
referred to as "barfoo" in other parts of the
PSX debug info, e.g.

	// register: 21
	register unsigned char barr;
	// register: 16
	register unsigned char barg;

Thus, we apply a similar naming convention for BarColor
and BarPos, using CamelCase as is conventional for global
variables in Diablo 1.

Updates #253.